refactor(particle): clarify orbital radial API
